--- beacon-chain/rpc/eth/node/server_test.go
@@ -1,7 +1,0 @@
-package node
-
-import (
-	ethpbservice "github.com/prysmaticlabs/prysm/v4/proto/eth/service"
-)
-
-var _ ethpbservice.BeaconNodeServer = (*Server)(nil)
